pyproject.toml
[tool]
[tool.poetry]
name = "muni-portal-backend"
version = "0.1.0"
description = ""
authors = ["JD Bothma <jd@openup.org.za>"]

[tool.poetry.dependencies]
python = "==3.8.x"
boto3 = "==1.16.2"
codecov = "==2.1.8"
django = "==2.2.10"
django-cors-headers = "==3.5.0"
django-environ = "==0.4.5"
django-extensions = "==2.2.6"
django-storages = "==1.10.1"
djangorestframework = "==3.11.1"
gevent = "==1.4.0"
gunicorn = "==20.0.4"
html5lib = "==1.0.1"
psycopg2 = "==2.8.4"
pytest-cov = "==2.8.1"
sentry-sdk = "==0.14.1"
wagtail = "==2.10.1"
werkzeug = "==0.16.0"
whitenoise = "==5.0.1"
django-rest-registration = "^0.5.6"
djangorestframework-simplejwt = "^4.5.0"
Faker = "^4.14.2"

[tool.poetry.dev-dependencies]
black = "==19.10b0"
isort = "==4.3.21"
